Backend/requirements.txt
bcrypt==3.2.0
fastapi==0.115.5
psycopg2==2.9.10
python-dotenv==1.0.1
python-jose==3.3.0
SQLAlchemy==2.0.36
starlette==0.41.2
uvicorn==0.32.0
pydantic==2.9.2
asyncpg==0.27.0  
